refactor: move cognito to own module
